+++ packages/map-template/releaseTools/webcomponent.js
@@ -1,7 +1,7 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import reactToWebComponent from 'react-to-webcomponent';
-import PropTypes from 'prop-types';
 import MapsIndoorsMap from '../src/components/MapsIndoorsMap/MapsIndoorsMap';
 
 /*
@@ -28,10 +28,12 @@ MapsIndoorsMap.propTypes = {
     pitch: PropTypes.number,
     kioskOriginLocationId: PropTypes.string,
     useMapProviderModule: PropTypes.bool,
-    timeout: PropTypes.number, 
+    timeout: PropTypes.number,
     miTransitionLevel: PropTypes.number,
+    category: PropTypes.string,
+    language: PropTypes.string,
 };
 
 const WebMapsIndoorsMap = reactToWebComponent(MapsIndoorsMap, React, ReactDOM);
-
+window.customElements.define('mapsindoors-map', WebMapsIndoorsMap);
 export default WebMapsIndoorsMap;