+++ packages/map-template/vite.config.js
@@ -1,8 +1,9 @@
-import { defineConfig } from 'vite';
+import { sentryVitePlugin } from "@sentry/vite-plugin";
 import react from '@vitejs/plugin-react';
-import svgr from 'vite-plugin-svgr';
+import { defineConfig } from 'vite';
 import eslint from 'vite-plugin-eslint';
 import { ViteFaviconsPlugin } from 'vite-plugin-favicon2';
+import svgr from 'vite-plugin-svgr';
 
 
 export default defineConfig(() => {
@@ -11,13 +12,22 @@ export default defineConfig(() => {
             port: 3000
         },
         build: {
-            outDir: 'build'
+            outDir: 'build',
+            sourcemap: true,
         },
         plugins: [
             react(),
             svgr(),
             ViteFaviconsPlugin('./public/favicon.png'),
-            eslint()
+            eslint(),
+            sentryVitePlugin({
+                org: process.env.SENTRY_ORG,
+                project: process.env.SENTRY_PROJECT,
+          
+                // Auth tokens can be obtained from https://sentry.io/orgredirect/organizations/:orgslug/settings/auth-tokens/
+                authToken: process.env.SENTRY_AUTH_TOKEN,
+                reactComponentAnnotation: { enabled: true },
+            }),
         ]
     }
 });