+++ packages/map-template/CHANGELOG.md
@@ -5,6 +5,42 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.41.0] - 2024-03-19
+
+### Added
+
+- Added Custom Element definition to Web Component generation file. This means you can avoid inline scripting, and only load a single JS-file from a CDN to get up and running with the Map Template Web Component
+
+## [1.40.0] - 2024-03-06
+
+### Added 
+
+- Added Sentry for Error Logging and Performance Metrics
+
+## [1.39.1] - 2024-03-04
+
+### Added 
+
+- Added docs for `category` prop.
+
+## [1.39.0] - 2024-02-29
+
+### Added 
+
+- Added legend modal to the kiosk.
+
+## [1.38.1] - 2024-02-28
+
+### Fixed 
+
+- Added missing `language` prop to web component.
+
+## [1.38.0] - 2024-02-28
+
+### Added 
+
+- Added support for `category` property.
+
 ## [1.37.1] - 2024-02-26
 
 ### Fixed